Invalidating associated change requests when a GeoObjectType is deleted.

+++ georegistry-server/src/main/java/net/geoprism/registry/service/ChangeRequestService.java
@@ -25,6 +25,8 @@ import org.commongeoregistry.adapter.action.AbstractActionDTO;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import com.runwaysdk.LocalizationFacade;
+import com.runwaysdk.dataaccess.ProgrammingErrorException;
 import com.runwaysdk.dataaccess.transaction.Transaction;
 import com.runwaysdk.query.OIterator;
 import com.runwaysdk.query.OrderBy.SortOrder;
@@ -38,6 +40,7 @@ import net.geoprism.registry.action.AbstractActionQuery;
 import net.geoprism.registry.action.AllGovernanceStatus;
 import net.geoprism.registry.action.ChangeRequest;
 import net.geoprism.registry.action.ChangeRequestQuery;
+import net.geoprism.registry.model.ServerGeoObjectType;
 
 public class ChangeRequestService
 {
@@ -279,6 +282,13 @@ public class ChangeRequestService
   public JSONObject deleteChangeRequest(String sessionId, String requestId)
   {
     ChangeRequest request = ChangeRequest.get(requestId);
+    
+    if (!request.isVisible())
+    {
+      // TODO : I want to throw CGRPermissionException but can't because it's in a different branch...
+      throw new ProgrammingErrorException("No permissions");
+    }
+    
     request.delete();
 
     return request.getDetails();
@@ -302,5 +312,26 @@ public class ChangeRequestService
     action.unlock();
 
     return action.serialize().toString();
+  }
+  
+  @Transaction
+  public void markAllAsInvalid(ServerGeoObjectType type)
+  {
+    String reason = LocalizationFacade.localize("changeRequest.invalidate.deleteReferencedGeoObjectType");
+    
+    ChangeRequestQuery crq = new ChangeRequestQuery(new QueryFactory());
+    
+    crq.WHERE(crq.getApprovalStatus().containsExactly(AllGovernanceStatus.PENDING));
+    
+    try (OIterator<? extends ChangeRequest> it = crq.getIterator())
+    {
+      for (ChangeRequest cr : it)
+      {
+        if (cr.referencesType(type))
+        {
+          cr.invalidate(reason);
+        }
+      }
+    }
   }
 }
